DFUtil.jl

Documentation for DFUtil.jl

DFUtil.sum_columns — Function

sum_columns(df, group_by::Vector{String}=Vector{String}())

Sum the columns of the dataframe, optionally grouping by group_by

All the columns not in the grouping must have a + method

Arguments

df DataFrame to sum

group_by Vector of String names of the columns to group by

replace_with by default DataFrames takes the column names and appends _function. Instead use this string.

DFUtil.group_data_into_periods — Function

group_data_into_periods(df::DataFrame, date_column::Union{Symbol, AbstractString}, period::Union{Symbol, AbstractString}; andgrpby = Vector{String}())

Using the given data_column, group the dataframe into periods.

All columns except the grouping ones must have methods for +`

Arguments

df DataFrame to group

date_column the column to use as the grouping

period Period to use, options are: :Quarter, :Month, :Year, :Week (or as strings)

andgrpby optional keyword to use additional groupings

Examples

group_data_into_periods(df, :SaleDate, :Week)
group_data_into_periods(df, "SaleDate", :Year, "BranchId")
group_data_into_periods(df, :SaleDate, "Quarter", [:Area, :BranchId)]

DFUtil.pQuarter — Function

pQuarter(dt)

turn a Date / Datetime into its eqivalent YearQn representation e.g. 2001Q1

DFUtil.pWeek — Function

pWeek(dt)

Turn a Date / Datetime into its eqivalent YearWeek representation e.g. 2001-01 Preserves 2000-01-01 becoming 1999-52

DFUtil.pYear — Function

pYear(dt)

turn a Date / Datetime into its eqivalent Year representation e.g. 2001

DFUtil.pMonth — Function

pMonth(dt)

turn a Date / Datetime into its eqivalent YearMonth representation e.g. 2001-01

DFUtil.match_row — Function

match_row(df, col, val)

Just a shortcut for filter(row -> row[col] == val, df)

DFUtil.to_json — Function

to_json(io::IO, data::DataFrame, keys::Union{AbstractString, Vector{AbstractString}})

Output the dataframe to Json, grouping by the given keys

{ "row[keys[1]]" : { "row[keys[2]]" : { "names(row)[1]" : "row[names(row[1])]", "names(row)[2]" : "row[names(row[2])]", ..., "names(row)[end]" : "row[names(row[end])]"}}}

The motivation for this is to enable referencing the objects via their nested keys in JavaScript

julia> df = DataFrame([[10, 10, 11], [1, 2, 3], ["Fred", "Luke", "Alice"]], ["Area", "EmpId", "Name"])
3×3 DataFrame
 Row │ Area   EmpId  Name
	 │ Int64  Int64  String
─────┼──────────────────────
   1 │    10      1  Fred
   2 │    10      2  Luke
   3 │    11      1  Alice

   julia> to_json(stdout, df, ["Area", "EmpId"])
   
   { "10" : {"1" : {"Area" : "10", "EmpId" : "1", "Name" : "Fred" } , "2" : {"Area" : "10", "EmpId" : "2", "Name" : "Luke" }  }, "11" : {"3" : {"Area" : "11", "EmpId" : "3", "Name" : "Alice" }  } }

in Javascript

let fred = object[10][1]	
let luke = object[10][2]		
let alice = object[11][1]

Arguments

io IO handle to write to

df DataFrame to write

keys Keys to lift outside the object

Examples

df = DataFrame([[1,2],[3,4],[4,5]],["a", "b", "c"]), "a")

to_json(stdout, df, "a")

"{ "1" : {"a" : "1", "b" : "3", "c" : "4" } , "2" : {"a" : "2", "b" : "4", "c" : "5" }  }"

to_json(stdout, df, ["a", "b"])

"{ "1" : {"3" : {"a" : "1", "b" : "3", "c" : "4" }  }, "2" : {"4" : {"a" : "2", "b" : "4", "c" : "5" }  } }")

DFUtil.to_json_var — Function

to_json_var(io::IO, df::DataFrame, keys::Union{AbstractString, Vector{AbstractString}}, var="object")

Output the dataframe to json using to_json but wrap it inside a Javascript JSON.parse() for direct use in Javascript

Arguments

io IO handle to write to

df DataFrame to write

keys Keys to lift outside the object - see to_json for explanation of that

var JavaScript object name it will be assign to with var $var = ....

Examples

df = DataFrame([[1,2],[3,4],[4,5]],["a", "b", "c"]), "a")

to_json_var(stdout, df, "a", "object")

"var object=JSON.parse('{ "1" : {"a" : "1", "b" : "3", "c" : "4" } , "2" : {"a" : "2", "b" : "4", "c" : "5" }  }');"

DFUtil.include_or_exclude — Function

include_or_exclude(df, includes, excludes)

Given a DataFrame either restrict it to the includes list, or remove the exludes list

Sounds a bit daft but it is to provide KW options in functions

If both are given, the includes take priority

Arguments

df the DataFrame

includes the String, Symbol or list of those to include

excludes the String, Symbol or list of those to exclude

Examples

df = DataFrame([[1,2,3], [10,20,30]], ["a", "b"])

julia> include_or_exclude(df, includes=["a"])
3×1 DataFrame
Row │ a
	│ Int64
─────┼───────
1 │     1
2 │     2
3 │     3

julia> include_or_exclude(df, excludes="a")
3×1 DataFrame
Row │ b
	│ Int64
─────┼───────
1 │    10
2 │    20
3 │    30

julia> include_or_exclude(df, includes=["a"], excludes=["a", "b"])
3×1 DataFrame
Row │ a
	│ Int64
─────┼───────
1 │     1
2 │     2
3 │     3

DFUtil.nthrow — Function

nthrow(df)

Return the nth row of a dataframe as a DataFrameRow, if n is out of bounds, return nothing

DFUtil.firstrow — Function

firstrow(df)

Return the first row as a DataFrameRow

DFUtil.lastrow — Function

lastrow(df)

Return the last row of df as a DataFrameRow

DFUtil.de_miss_rows — Function

de_miss_rows(df)

For a given DataFrame, remove any rows in which any column has a missing value

Examples

df = DataFrame([[1,missing,3], [10,20,30]], ["a", "b"])
3×2 DataFrame
 Row │ a        b
	 │ Int64?   Int64?
─────┼─────────────────
   1 │       1      10
   2 │ missing      20
   3 │       3      30
	
julia> dm = de_miss_rows(df)
2×2 DataFrame
 Row │ a       b
	 │ Int64?  Int64?
─────┼────────────────
   1 │      1      10
   2 │      3      30

DFUtil.to_csv_text — Function

to_csv_text(df)

turn the dataframe into the string representation of the CSV

Examples

df = DataFrame(
[[53.685335, 53.785335, 53.885335],
[-0.416387,-0.426387,-0.436387],
[53.479544, 53.479544,53.479544],
[-2.954101,-2.954101,-2.954101]], ["lat1", "lon1", "lat2", "lon2"])

to_csv_text(df)
"lat1,lon1,lat2,lon2\n53.685335,-0.416387,53.479544,-2.954101\n53.785335,-0.426387,53.479544,-2.954101\n53.885335,-0.436387,53.479544,-2.954101\n"

DFUtil.from_csv_text — Function

from_csv_text(csv_text)

Create a dataframe from the string represenation of a CSV

Examples

from_csv_text("lat1,lon1,lat2,lon2\n53.685335,-0.416387,53.479544,-2.954101\n53.785335,-0.426387,53.479544,-2.954101\n53.885335,-0.436387,53.479544,-2.954101\n")

3×4 DataFrame
Row │ lat1     lon1       lat2     lon2
	│ Float64  Float64    Float64  Float64

─────┼────────────────────────────────────── 1 │ 53.6853 -0.416387 53.4795 -2.9541 2 │ 53.7853 -0.426387 53.4795 -2.9541 3 │ 53.8853 -0.436387 53.4795 -2.9541