DFUtil.jl

Documentation for DFUtil.jl

DFUtil.sum_columns — Function

sum_columns(df, group_by::Vector{String}=Vector{String}())

sum the columns of the dataframe, optionally grouping by group_by

Arguments

df DataFrame to sum

group_by Vector of String names of the columns to group by

replace_with by default DataFrames takes the column names and appends _function. Instead use this string.

DFUtil.group_data_into_periods — Function

group_data_into_periods(df::DataFrame, date_column::Union{Symbol, AbstractString}, period::Union{Symbol, AbstractString}; andgrpby = Vector{String}())

Using the given data_column, group the dataframe into periods.

All columns except the grouping ones *must* have methods for +

Arguments

df DataFrame to group

date_column the column to use as the grouping

period Period to use, options are: :Qtr, :Month, :Year, :Week (or as strings)

andgrpby optional keyword to use additional groupings

Examples

`group_data_into_periods(df, :SaleDate, :Week)`
`group_data_into_periods(df, "SaleDate", :Year, "BranchId")`
`group_data_into_periods(df, :SaleDate, "Qtr", [:Area, :BranchId)]`

DFUtil.match_row — Function

match_row(df, col, val)

Just a shortcut for filter(row -> row[col] == val, df)

Not even sure why I introduced it

DFUtil.to_json — Function

to_json(io::IO, data::DataFrame, keys::Union{AbstractString, Vector{AbstractString}})

Output the dataframe to Json, grouping by the given keys

`{ "row[keys[1]]" : { "row[keys[2]]" : { "names(row)[1]" : "row[names(row[1])]", "names(row)[2]" : "row[names(row[2])]", ..., "names(row)[end]" : "row[names(row[end])]"}}}`

Arguments

io IO handle to write to

df DataFrame to write

keys Keys to lift outside the object

DFUtil.to_json_var — Function

to_json_var(io::IO, df::DataFrame, keys::Union{AbstractString, Vector{AbstractString}}, var="object")

Output the dataframe to json using to_json but wrap it inside a Javascript JSON.parse() for direct use in Javascript

Arguments

io IO handle to write to

df DataFrame to write

keys Keys to lift outside the object - see to_json for explanation of that

var JavaScript object name it will be assign to with var $var = ....